backend/requirements.txt
# Core
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Firebase & GCP
firebase-admin==6.4.0
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-cloud-bigquery==3.17.2
google-cloud-logging==3.9.0

# Redis
redis==5.0.1
hiredis==2.3.2

# HTTP & Async
httpx==0.26.0
aiohttp==3.9.3

# AI/ML
openai==1.12.0
pinecone-client==3.0.3
tiktoken==0.6.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Testing
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code Quality
black==23.12.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2

# Utilities
structlog==24.1.0
tenacity==8.2.3

